# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arjuno Travel API",
        "description": "Arjuno Travel Booking Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/travel": {
            "get": {
                "tags": [
                    "Travel"
                ],
                "summary": "Get a list of available travels",
                "operationId": "12a0f892a4961c4dff264000058506d4",
                "parameters": [
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Filter by price",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "Filter by origin",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "Filter by destination",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "departure_time",
                        "in": "query",
                        "description": "Filter by departure time",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Travel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Travel"
                ],
                "summary": "Add a new travel",
                "operationId": "efb2afdd4dd517ca54be0004f2136258",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddTravelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Travel added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/travel/{id}": {
            "get": {
                "tags": [
                    "Travel"
                ],
                "summary": "Get details of a specific travel",
                "operationId": "fca87f02f34abc66255db00d4ee61ff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Travel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "put": {
                "tags": [
                    "Travel"
                ],
                "summary": "Update a travel",
                "operationId": "1ffb400d4da13109d9e05eb8f4fb839b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddTravelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Travel updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Travel"
                ],
                "summary": "Delete a travel",
                "operationId": "5f9c5555c53530685f1bfe539ead5f5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the travel to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Travel deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden. You are not authorized to delete this travel"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Travel": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "origin": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-21 12:00:00"
                    }
                },
                "type": "object"
            },
            "AddTravelRequest": {
                "required": [
                    "price",
                    "origin",
                    "destination",
                    "departure_time"
                ],
                "properties": {
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "origin": {
                        "type": "string"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "is_available": {
                        "type": "boolean"
                    },
                    "departure_time": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-07-21 12:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}